chore: refactor common code
